[REF] web: move pdfjs utils to core utils
