+++ addons/mail/__manifest__.py
@@ -221,7 +221,6 @@ For more specific needs, you may also assign custom-defined actions
             'web/static/src/scss/mimetypes.scss',
             'web/static/src/scss/ui.scss',
             ('include', 'web._assets_core'),
-            'web/static/src/libs/pdfjs.js',
             'web/static/src/views/fields/formatters.js',
             'web/static/src/views/fields/file_handler.*',
