+++ addons/web/__manifest__.py
@@ -115,8 +115,6 @@ This module provides the core of the Odoo Web Client.
             'web/static/src/webclient/actions/reports/*.js',
             'web/static/src/webclient/actions/reports/*.xml',
 
-            'web/static/src/libs/pdfjs.js',
-
             'web/static/src/scss/ace.scss',
             'web/static/src/scss/base_document_layout.scss',
 
@@ -238,7 +236,6 @@ This module provides the core of the Odoo Web Client.
             'web/static/src/public/error_notifications.js',
             'web/static/src/public/public_component_service.js',
             'web/static/src/public/datetime_picker_widget.js',
-            'web/static/src/libs/pdfjs.js',
 
             'web/static/src/legacy/js/public/public_root.js',
             'web/static/src/legacy/js/public/public_root_instance.js',
